+++ packages/forma-36-react-components/src/components/Flex/Flex.tsx
@@ -5,7 +5,17 @@ import cn from 'classnames';
 
 import styles from './Flex.css';
 
-export type SpacingTypes = 'none' | 'spacing2xs' | 'spacingXs' | 'spacingS' | 'spacingM' | 'spacingL' | 'spacingXl' | 'spacing2Xl' | 'spacing3Xl' | 'spacing4Xl';
+export type SpacingTypes =
+  | 'none'
+  | 'spacing2xs'
+  | 'spacingXs'
+  | 'spacingS'
+  | 'spacingM'
+  | 'spacingL'
+  | 'spacingXl'
+  | 'spacing2Xl'
+  | 'spacing3Xl'
+  | 'spacing4Xl';
 
 export interface FlexProps {
   /**
@@ -20,7 +30,7 @@ export interface FlexProps {
   /**
    * Html tag used as container
    */
-  htmlTag?: any,
+  htmlTag?: React.ElementType;
   /**
    * Sets width: 100% */
   fullWidth?: boolean;
@@ -35,14 +45,17 @@ export interface FlexProps {
   noShrink?: boolean;
   /**
    * One of flex-wrap css values */
-  flexWrap?: CSS.FlexWrapProperty,
+  flexWrap?: CSS.FlexWrapProperty;
   /**
    * One of flex-direction css values */
   flexDirection?: CSS.FlexDirectionProperty;
   /**
+   * One of flex-grow css values */
+  flexGrow?: CSS.GlobalsNumber;
+  /**
    * One of justify-content css values */
   justifyContent?: CSS.JustifyContentProperty;
-   /**
+  /**
    * One of justify-content css values */
   justifyItems?: CSS.JustifyContentProperty;
   /**
@@ -89,7 +102,8 @@ export interface FlexProps {
   style?: React.CSSProperties;
 }
 
-const handleSpacing = (value: SpacingTypes) => value === 'none' ? 0 : tokens[value];
+const handleSpacing = (value: SpacingTypes) =>
+  value === 'none' ? 0 : tokens[value];
 
 const defaultProps = {
   testId: 'cf-ui-flex',
@@ -103,6 +117,7 @@ export const Flex = (props: FlexProps) => {
     noShrink,
     flexWrap,
     flexDirection,
+    flexGrow,
     inlineFlex,
     justifyContent,
     justifyItems,
@@ -126,21 +141,21 @@ export const Flex = (props: FlexProps) => {
     ...otherProps
   } = props;
 
-  const fullMargins = {margin: margin && handleSpacing(margin)}
+  const fullMargins = { margin: margin && handleSpacing(margin) };
   const sidesMargins = {
     marginTop: marginTop && handleSpacing(marginTop),
     marginRight: marginRight && handleSpacing(marginRight),
     marginBottom: marginBottom && handleSpacing(marginBottom),
     marginLeft: marginLeft && handleSpacing(marginLeft),
-  }
+  };
 
-  const fullPaddings = {padding: padding && handleSpacing(padding)}
+  const fullPaddings = { padding: padding && handleSpacing(padding) };
   const sidesPaddings = {
     paddingTop: paddingTop && handleSpacing(paddingTop),
     paddingRight: paddingRight && handleSpacing(paddingRight),
     paddingBottom: paddingBottom && handleSpacing(paddingBottom),
     paddingLeft: paddingLeft && handleSpacing(paddingLeft),
-  }
+  };
   const marginResult = margin ? fullMargins : sidesMargins;
   const paddingsResult = padding ? fullPaddings : sidesPaddings;
 
@@ -152,7 +167,8 @@ export const Flex = (props: FlexProps) => {
     alignItems,
     alignSelf,
     flexWrap,
-  }
+    flexGrow,
+  };
 
   const classNames = cn(styles.Flex, className, {
     [styles.Flex__inline]: inlineFlex,
@@ -161,13 +177,17 @@ export const Flex = (props: FlexProps) => {
     [styles.Flex__noShrink]: noShrink,
   });
 
-
   return (
-    <Container {...otherProps} style={{...inlineStyle, ...marginResult, ...paddingsResult, ...style}} className={classNames} data-test-id={testId}>
+    <Container
+      {...otherProps}
+      style={{ ...inlineStyle, ...marginResult, ...paddingsResult, ...style }}
+      className={classNames}
+      data-test-id={testId}
+    >
       {children}
     </Container>
   );
-}
+};
 Flex.defaultProps = defaultProps;
 
 export default Flex;