+++ packages/forma-36-tokens/src/tokens/z-index.js
@@ -1,13 +1,13 @@
 const zIndex = {
-  'z-index-negative': '-1',
-  'z-index-workbench': '0',
-  'z-index-default': '1',
-  'z-index-workbench-header': '10',
-  'z-index-modal': '100',
-  'z-index-modal-content': '101',
-  'z-index-dropdown': '1000',
-  'z-index-tooltip': '10000',
-  'z-index-notification': '100000',
+  'z-index-negative': -1,
+  'z-index-workbench': 0,
+  'z-index-default': 1,
+  'z-index-workbench-header': 10,
+  'z-index-modal': 100,
+  'z-index-modal-content': 101,
+  'z-index-dropdown': 1000,
+  'z-index-tooltip': 10000,
+  'z-index-notification': 100000,
 };
 
 module.exports = zIndex;